Code readability improvements.
